move types to pkg
