+++ cmd/registry/scoring/score.go
@@ -21,9 +21,9 @@ import (
 	"strings"
 
 	"github.com/apigee/registry/cmd/registry/patterns"
-	"github.com/apigee/registry/cmd/registry/types"
 	"github.com/apigee/registry/pkg/log"
 	"github.com/apigee/registry/pkg/names"
+	"github.com/apigee/registry/pkg/types"
 	"github.com/apigee/registry/rpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
